fastapi-backend/requirements.txt
fastapi
uvicorn
torch
openai-whisper
python-multipart
